Support inline examples as fuzzable values
This change introduces the following behavior:

- If example values are specified in the Swagger specification, they
are plugged into default values in the grammar (instead of 'fuzzstring', etc.).
These do not affect the schema.

- If example payloads are specified outside Swagger, these are used
in the same way as before - both to determine the schema and to select values.

The two are mutually exclusive.  For example, if 'useQueryExamples' is specified,
and a query example is found in external payload examples, they are used and
the Swagger spec values are ignored.